[clang] warnings cleanup

+++ win32ss/user/ntuser/winsta.h
@@ -30,8 +30,8 @@ typedef struct _WINSTATION_OBJECT
     DWORD          cNumClipFormats;
     INT            iClipSerialNumber;
     INT            iClipSequenceNumber;
-    INT            fClipboardChanged : 1;
-    INT            fInDelayedRendering : 1;
+    ULONG          fClipboardChanged : 1;
+    ULONG          fInDelayedRendering : 1;
 
     PWND           spwndClipboardListener;
     LUID           luidEndSession;
